refactor : 기존 세션 이미지에서 삭제할 이미지가 없는 경우에 대한 조건 추가

+++ src/main/java/com/epris/homepage/activity/session/service/SessionService.java
@@ -49,7 +49,7 @@ public class SessionService {
             }
         }
         /* 기존 이미지 삭제*/
-        deleteSessionImageList(deleteSessionImageList);
+        if(!deleteSessionImageList.isEmpty())deleteSessionImageList(deleteSessionImageList);
 
         /* 세션 업데이트 */
         updateSession.update(requestDto.getSessionInfo());
